add models and IRegionInterface

+++ Models/FactsToRegion.cs
@@ -3,7 +3,7 @@ using System;
 namespace chronos.Models{
     public class FactsToRegion
     {
-        public int fact_id { get; set; }
-        public int region_id { get; set; } 
+        public long fact_id { get; set; }
+        public long region_id { get; set; } 
     }
 }
